+++ src/ni_measurement_plugin_converter/models/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 
 from ni_measurement_plugin_converter.models._cli import CliInputs, InvalidCliArgsError
+from ni_measurement_plugin_converter.models._exceptions import UnsupportedDriverError
 from ni_measurement_plugin_converter.models._inputs_outputs import InputInfo, OutputInfo