+++ src/ni_measurement_plugin_converter/models/_cli.py
@@ -7,7 +7,6 @@ from typing import Optional
 from pydantic import BaseModel, model_validator
 
 from ni_measurement_plugin_converter.constants import ENCODING, UserMessage
-
 from ._exceptions import InvalidCliArgsError